Update index.md
HTML is not a programming language. It is a markup language, and its full name is Hyper Text Markup Language. 
HTML is used to create the basic structure of a website. It's easy to learn compared to programming languages, but you can't do much with it and you can't program with it.
Cascading Style Sheets (CSS) is a stylesheet language that instructs the browser how we want our HTML to appear; it is not a programming language.